+++ Engine/WriteNode.cpp
@@ -328,13 +328,13 @@ WriteNodePrivate::placeWriteNodeKnobsInPage()
     }
     if (foundSep != -1) {
         ++foundSep;
-        if (foundSep < children.size()) {
+        if (foundSep < (int)children.size()) {
             bool isSecret = children[foundSep]->getIsSecret();
-            while (isSecret && foundSep < children.size()) {
+            while (isSecret && (int)foundSep < children.size()) {
                 ++foundSep;
                 isSecret = children[foundSep]->getIsSecret();
             }
-            if (foundSep < children.size()) {
+            if (foundSep < (int)children.size()) {
                 separatorKnob.lock()->setSecret(dynamic_cast<KnobSeparator*>(children[foundSep].get()));
             } else {
                 separatorKnob.lock()->setSecret(true);
